+++ example/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -1,0 +1,12 @@
+//
+//  Generated file. Do not edit.
+//
+
+import FlutterMacOS
+import Foundation
+
+import flui_kit
+
+func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
+  FluiKitPlugin.register(with: registry.registrar(forPlugin: "FluiKitPlugin"))
+}