+++ lib/librairy.dart
@@ -1,0 +1,1 @@
+export 'src/widgets/plan_switcher.dart';